requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
opencv-python
pillow
numpy
tqdm
python-multipart
python-dotenv
